Add files via upload
I use numpy to calculate the gravitational acceleration on a hypothetical object at a distance from each planet in the solar system which is input by the user. It does this by running column operations on the included spreadsheet which has the mass and radius of each planet. The results are appended to the spreadsheet.